+++ react/MainCategories.js
@@ -1,0 +1,3 @@
+import MainCategories from './components/MainCategories'
+
+export default MainCategories